requirements-lock.txt
# Auto-generated lock file — DO NOT EDIT MANUALLY
# Generated from: pip freeze (venv_win, Python 3.10.11, Windows 10)
# Date: 2026-02-15
# Purpose: Reproducible builds with exact dependency versions
#
# To regenerate: pip freeze > requirements-lock.txt
# To install:    pip install -r requirements-lock.txt
#
# Note: torch packages require the PyTorch index for CUDA builds.
# For CPU-only (CI), use requirements-ci.txt instead.

--extra-index-url https://download.pytorch.org/whl/cu121

absl-py==2.3.1
aiofiles==25.1.0
alembic==1.18.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
asyncgui==0.6.3
asynckivy==0.6.4
awpy==1.2.3
backports.tarfile==1.2.0
beautifulsoup4==4.14.3
build==1.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
coverage==7.13.0
cryptography==46.0.3
cycler==0.12.1
Cython==0.29.37
demoparser2==0.40.2
distlib==0.4.0
docutils==0.22.4
etils==1.13.0
exceptiongroup==1.3.1
fastapi==0.128.0
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.0.2
jaraco.functools==4.4.0
Jinja2==3.1.6
joblib==1.5.3
keyring==25.6.0
Kivy==2.3.0
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip#sha256=033613238896799a6f7bb3fc451d2d8191515b3b8a0ea1e04682599b19dcf83e
kiwisolver==1.4.9
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==2.1.5
materialshapes==0.3
materialyoucolor==2.0.10
matplotlib==3.10.8
mdurl==0.1.2
meson==1.10.0
ml_dtypes==0.5.4
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
ncps==1.0.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
opencv-python==4.13.0.92
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pefile==2024.8.26
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
pluggy==1.6.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pyarrow==22.0.0
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
PyMuPDF==1.26.7
pyparsing==3.3.1
pypdf==6.7.0
pypdfium2==5.4.0
pypiwin32==223
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.52.0
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
sympy==1.13.1
textdistance==4.6.3
threadpoolctl==3.6.0
tomli==2.3.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
treescope==0.1.10
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
virtualenv==20.35.4
zipp==23.0
